pyproject.toml
[project]
name = "loyverse-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.2.0",
    "httpx>=0.28.1",
    "ipython>=8.37.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.5",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
]
